client side of sprint 2

+++ app/Http/Controllers/OrderController.php
@@ -6,10 +6,12 @@ use App\Http\Requests\StoreOrderRequest;
 use App\Http\Requests\UpdateOrderRequest;
 use App\Http\Resources\OrderResource;
 use App\Models\Bottle;
+use App\Models\bottle_order;
 use App\Models\BottleOrder;
 use App\Models\Client;
 use App\Models\Order;
 use App\Models\Perfume;
+use App\Models\User;
 use DB;
 use Exception;
 
@@ -18,7 +20,10 @@ class OrderController extends Controller
     public function __construct(
         private Client $client,
         private Order $order,
-        private BottleOrder $bottleOrder
+        private User $user,
+        private BottleOrder $bottleOrder,
+        private Perfume $perfume,
+        private bottle_order $bottle_order,
     ) {
     }
 
@@ -51,9 +56,9 @@ class OrderController extends Controller
                 return $this->sendError('Client not found', '', 404);
             }
 
-            $perfume = Perfume::find($request->input('perfume'));
+            $perfume = Perfume::find($request->input('perfume_id'));
             if (! $perfume) {
-                return $this->sendError('Perfume with ID '.$request->input('perfume').' not found', '', 404);
+                return $this->sendError('Perfume with ID '.$request->input('perfume_id').' not found', '', 404);
             }
 
             // New order
@@ -84,7 +89,7 @@ class OrderController extends Controller
 
             return $this->sendResponse('order placed', new OrderResource($order));
 
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             DB::rollBack();
             // Handle the exception and return an error response
             return $this->sendError('Error occurred while storing the order', $e->getMessage(), 500);
@@ -99,7 +104,7 @@ class OrderController extends Controller
             return $this->sendError('client not found', '', 404);
         }
 
-        return $this->sendResponse('', (OrderResource::collection($client->orders()->get())));
+        return $this->sendResponse('ok', (OrderResource::collection($client->orders()->get())));
 
     }
 
@@ -142,12 +147,25 @@ class OrderController extends Controller
             if (! $order) {
                 return $this->sendError('order not found', '', 404);
             }
-            $order->bottles()->detach($bottle_id);
+
+            $pivot = $this->bottle_order::where('bottle_id', $bottle_id)->where('order_id', $order_id)->first();
+
+            if ($pivot) {
+                if ($pivot->quantity === 1) {
+                    $order->bottles()->detach($bottle_id);
+                } else {
+                    DB::table('bottle_order')
+                        ->where('order_id', $order_id)
+                        ->where('bottle_id', $bottle_id)
+                        ->update(['quantity' => $pivot->quantity - 1]);
+
+                }
+            }
 
             return $this->sendResponse('deleted');
 
         } catch (Exception $e) {
-            //throw $th;
+            return $this->sendError('error', $e->getMessage(), 500);
         }
     }
 
@@ -196,5 +214,65 @@ class OrderController extends Controller
             // Handle error if order is not found
             return $this->sendError('Order not found', '', 404);
         }
+    }
+
+    public function get_by_client($user_id)
+    {
+        $user = $this->user::find($user_id);
+        if (! $user) {
+            return $this->sendError('User not found', '', 404);
+        }
+
+        $client = $user->person;
+        if (! $client) {
+            return $this->sendError('client not found', '', 404);
+        }
+
+        $orders = $client->orders()->get();
+        $orders_coll = [];
+
+        foreach ($orders as $order) {
+            $total_price = 0;
+            $perfume = $this->perfume::find($order->perfume_id);
+            if (! $perfume) {
+                return $this->sendError('perfume not found', '', 404);
+            }
+
+            $bottlesIds = [];
+            // $bottles_status = [];
+            // $bottles_quant = [];
+            foreach ($order->bottles as $bottle) {
+
+                $quantity = $bottle->pivot->quantity;
+                $status = $bottle->pivot->status;
+                // dd($quantity);
+                $total_price += ($bottle->price + (($bottle->price * $perfume->extra_price) / 100)) * $quantity;
+                array_push($bottlesIds, $bottle->id);
+                // array_push($bottles_quant, $quantity);
+                // array_push($bottles_status, $status);
+                // for ($i = 0; $i < $quantity; $i++) {
+                // }
+            }
+
+            array_push($orders_coll, [
+                'id' => $order->id,
+                'perfume_id' => $perfume->id,
+                'status' => $order->status,
+                'total_price' => $total_price,
+                'bottles' => $bottlesIds,
+                // 'bottles_quantities' => $bottles_quant,
+                // 'bottles_status' => $bottles_status,
+            ]);
+        }
+        // dd($orders_coll);
+
+        return $this->sendResponse('orders retrieved', $orders_coll);
+        /**
+         * {
+         *  perfume_id -> integer,
+         *  bottles -> array of integers
+         *  total_price -> double
+         *  status
+         * } */
     }
 }
